requirements.txt
aiogram==3.10.0
Flask[async]==3.0.3
python-dotenv==1.0.1
SQLAlchemy==2.0.31
aiosqlite==0.20.0
